% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_snippets.R
\name{update_texsnippets}
\alias{update_texsnippets}
\title{update texsnippets}
\usage{
update_texsnippets()
}
\value{
A message about the success of the copy.
}
\description{
Update snippets in ~/.R/snippets/r.snippets using current package snippets
}
\examples{
\dontrun{
update_texsnippets()
}

}
