% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_snippets.R
\name{use_texsnippets}
\alias{use_texsnippets}
\title{setup texsnippets}
\usage{
use_texsnippets()
}
\value{
A message about the success of the copy.
}
\description{
Add the snippets to your ~/.R/snippets/r.snippets
}
\examples{
\dontrun{
use_texsnippets()
}
}
